+++ R/utils_measures.R
@@ -267,7 +267,7 @@ get_measures <- function(object,
                 })
         )
       names(smr) <- c("n", "area_sum", "area_mean", "area_sd",  names(res[6:ncol(res)]))
-      smr$img <- object$count$Image
+      smr$img <- object$count$img
       smr <- smr[,c(ncol(smr), 1:ncol(smr)-1)]
       smr$area_sd[is.na(smr$area_sd)] <- 0
       merg <- smr